Start implementing KeyFinder

+++ src/main/java/de/andidebob/language/BiGram.java
@@ -1,4 +1,4 @@
 package de.andidebob.language;
 
-public record BiGram(String raw, float frequency) {
+public record BiGram(String raw, double frequency) {
 }
